Add #each_page_with_cursor method to Enumerable
This is a reimplementation of what we did originally in 2014. See here:
https://github.com/sferik/twitter/pull/626

For the past 2 years we've been using an old fork of the Twitter gem.

The problem today is, in 2016 the Twitter API has changed and this old version
of the gem has issues. Specifically, it's ability to upload/tweet media
(gifs, multiple images).

So, this commit, re-implements the changes we had in 2014 (with paging
and exposing the cursor/rate limit). Just slightly different because the gem as
changed so much in that time.

We do risk this happening to us again in the future. But we consider
that risk worth it since hitting the rate limit often can result in
Twitter apps being banned, which would be a disaster for us.